Fix infinite loop when parsing invalid DAP message.
This commit reworks `ContentReader::read` to better validate
the incoming DAP message, closing the underlying `dap::Reader` in
the event of an invalid message.

This fixes an infinite loop when parsing an invalid DAP message. In
particular, the infinite loop occurs when a cross-site HTTP request from
a browser running on the same host as a cppdap server is sent to the
server. As cross-site requests to the localhost can be triggered by
untrusted JavaScript, more validation of the DAP messages is required to
protect cppdap-based servers from malicious requests.

This commit also includes:
- Changes to `ContentReader::read` to block until the complete message
  has been read from the underlying `dap::Reader` instead of returning
  an incomplete result when the underlying `dap::Reader` does not
  have data.
- A few new unit tests to test invalid DAP messages.
- Some changes to the existing unit tests for `ContentReader`, in
  particular those that were testing partial reads and discarding
  invalid messages without expecting the reader to be closed after
  receiving an invalid message.